requirements.txt
asgiref==3.6.0
boto3==1.26.59
botocore==1.29.59
Django==4.1.5
django-secrets-manager==0.1.13
jmespath==1.0.1
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
